feat(gsoc): update prompt component structure to make it a common helper component
